Refactored package TownyChat.event to TownyChat.listener
