+++ src/com/palmergames/bukkit/TownyChat/Chat.java
@@ -13,8 +13,8 @@ import com.palmergames.bukkit.TownyChat.CraftIRCHandler;
 import com.palmergames.bukkit.TownyChat.Command.TownyChatCommand;
 import com.palmergames.bukkit.TownyChat.channels.ChannelsHolder;
 import com.palmergames.bukkit.TownyChat.config.ConfigurationHandler;
-import com.palmergames.bukkit.TownyChat.event.HeroicDeathForwarder;
-import com.palmergames.bukkit.TownyChat.event.TownyPlayerHighestListener;
+import com.palmergames.bukkit.TownyChat.listener.HeroicDeathForwarder;
+import com.palmergames.bukkit.TownyChat.listener.TownyPlayerHighestListener;
 import com.palmergames.bukkit.TownyChat.tasks.onLoadedTask;
 import com.palmergames.bukkit.TownyChat.util.FileMgmt;
 import com.palmergames.bukkit.towny.Towny;
