+++ src/com/palmergames/bukkit/TownyChat/listener/TownyPlayerHighestListener.java
@@ -1,4 +1,4 @@
-package com.palmergames.bukkit.TownyChat.event;
+package com.palmergames.bukkit.TownyChat.listener;
 
 import java.util.WeakHashMap;
 
@@ -14,7 +14,7 @@ import com.palmergames.bukkit.towny.TownyMessaging;
 import com.palmergames.bukkit.TownyChat.channels.Channel;
 import com.palmergames.bukkit.TownyChat.channels.channelTypes;
 import com.palmergames.bukkit.TownyChat.config.ChatSettings;
-import com.palmergames.bukkit.TownyChat.event.TownyChatEvent;
+import com.palmergames.bukkit.TownyChat.listener.TownyChatEvent;
 import com.palmergames.bukkit.TownyChat.Chat;
 import com.palmergames.bukkit.TownyChat.TownyChatFormatter;
 import com.palmergames.bukkit.towny.exceptions.NotRegisteredException;
